Big refactoring of the AzureRefreshService class and the way we get catalogue info (VM Sizes, VM OS and Region objects) from Azure. Now the process makes sure to take into account the hierarchy so that the database can reflect which info is available per region.
Database changes: Modified some of the pre-seeded data and added two new tables to record the region-to-resource mappings

Web and app configs: Added support to target a specific region or publisher to reduce the amount of time the extension needs to query Azure and get said info.

Other minor edits.