The timeout issue is fixed with full live Azure functionality
Good: The timeout issue is fixed with full live Azure functionality
Bad: I put some amount of code in the controller which really belongs in
libraries. It is ugly and bad architecture. It will get us unblocked but
I believe we should still get it working again as it intended to work in
the libs. Not as pri 1, maybe pri 2 and not blocking release.