Temp fix in CmpDb.cs when using KryptoLib in a loop.
Turns out that we dispose of the PrivateKey as suggested by MSDN, but this renders the Krypto instance useless for further decrypts.

For now, we'll just create a new Krypto instance to use for each decrypt in that code block. We will need to refactor the Krypto class a bit to account for this and do the dispose in a better way.